Under test (#1)
* under_test
一键录入上次信息
异常逻辑优化

* 优化逻辑，直接显示上次记录

* 增加离发车时间近抢票后的确定提示操作

* implicitly_wait updates to 8s
optimize refresh tickets exception handling

* optimize login page logic

* module name update
