+++ ticket_with_autologin/info.dir
@@ -5,4 +5,4 @@
 'train', (2048, 15)
 'who', (2560, 19)
 'seat', (3072, 19)
-'driver', (3584, 16)
+'driver', (3584, 17)
